+++ dev/bots/test/analyze-test-input/root/packages/foo/flaky_golden_missing_tag.dart
@@ -2,9 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// This would fail analysis, but it is ignored
-// flutter_ignore_for_file: golden_tag (see analyze.dart)
-
+// The reduced test set tag is missing. This should fail analysis.
 @Tags(<String>['some-other-tag'])
 
 import 'package:test/test.dart';
@@ -12,6 +10,5 @@ import 'package:test/test.dart';
 import 'golden_class.dart';
 
 void main() {
-  matchesGoldenFile('missing_tag.png');
-  expectFlakyGolden('key', 'String');
+  expectFlakyGolden('finder', 'missing_tag.png');
 }
