+++ dev/bots/test/analyze-test-input/root/packages/foo/flaky_golden_no_tag.dart
@@ -1,0 +1,11 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+// The tag is missing. This should fail analysis.
+
+import 'golden_class.dart';
+
+void main() {
+  expectFlakyGolden('key', 'missing_tag.png');
+}
